+++ src/main/java/com/opentable/db/postgres/embedded/BundledPostgresBinaryResolver.java
@@ -28,4 +28,16 @@ final class BundledPostgresBinaryResolver implements PgBinaryResolver {
         return EmbeddedPostgres.class.getResourceAsStream(format("/postgresql-%s-%s.txz", system, machineHardware));
     }
 
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        return o != null && getClass() == o.getClass();
+    }
+
+    @Override
+    public int hashCode() {
+        return getClass().hashCode();
+    }
 }